requirements.txt
dkpro-cassis==0.2.7 
nltk==3.4.4
spacy==2.1.7
notebook==6.0.3 